requirements.txt
streamlit
sentence-transformers
transformers
torch
faiss-cpu
PyPDF2
accelerate
